Reorganize component system
Small improvements to the component system:
- Components in their own file
- Renamed `content` to `body` for clarity
- Moved generation logic to `tracker.ts` file
